Move ImageStorage from storage to commons package and rename to ImageUtil
ImageStorage does not fit the interface of storage. Therefore, it is moved away from storage to avoid confusion.
