+++ src/main/java/seedu/address/ui/PersonCard.java
@@ -14,7 +14,7 @@ import javafx.scene.layout.HBox;
 import javafx.scene.layout.Region;
 import seedu.address.model.person.Person;
 import seedu.address.model.tag.Tag;
-import seedu.address.storage.ImageStorage;
+import seedu.address.commons.util.ImageUtil;
 
 /**
  * An UI component that displays information of a {@code Person}.
@@ -63,7 +63,7 @@ public class PersonCard extends UiPart<Region> {
         id.setText(displayedIndex + ". ");
         name.setText(person.getName().fullName);
         // Add gender icon
-        Image genderImage = ImageStorage.getGenderIcon(person.getGender().gender);
+        Image genderImage = ImageUtil.getGenderIcon(person.getGender().gender);
         if (genderImage != null) {
             ImageView genderIcon = new ImageView(genderImage);
             genderIcon.setFitHeight(15);
@@ -107,7 +107,7 @@ public class PersonCard extends UiPart<Region> {
                 .sorted(Comparator.comparing(socialHandle -> socialHandle.platform))
                 .forEach(socialHandle -> {
                     Label label = new Label();
-                    Image image = ImageStorage.getSocialIcon(socialHandle.platform);
+                    Image image = ImageUtil.getSocialIcon(socialHandle.platform);
                     if (image == null) {
                         label.setText(socialHandle.platform + " : " + socialHandle.value);
                     } else {
